chore, refactor: crear una clase y lo que conlleva para las operaciones realizadas

+++ app/src/main/java/dev/wdona/gestorinventarioqr/data/api/EstanteriaApi.java
@@ -3,5 +3,5 @@ package dev.wdona.gestorinventarioqr.data.api;
 import dev.wdona.gestorinventarioqr.domain.model.Estanteria;
 
 public interface EstanteriaApi {
-    Estanteria getEstanteriaById(int id);
+    Estanteria getEstanteriaById(Long id);
 }
