+++ app/src/main/java/dev/wdona/gestorinventarioqr/data/datasource/remote/impl/EstanteriaRemoteDataSourceImpl.java
@@ -10,7 +10,7 @@ public class EstanteriaRemoteDataSourceImpl {
         this.api = api;
     }
 
-    public Estanteria getEstanteriaById(int id) {
+    public Estanteria getEstanteriaById(Long id) {
         return api.getEstanteriaById(id);
     }
 }
