+++ app/src/main/java/dev/wdona/gestorinventarioqr/data/repository/EstanteriaRepository.java
@@ -3,5 +3,6 @@ package dev.wdona.gestorinventarioqr.data.repository;
 import dev.wdona.gestorinventarioqr.domain.model.Estanteria;
 
 public interface EstanteriaRepository {
-        Estanteria getEstanteriaById(int id);
+
+        Estanteria getEstanteriaById(Long id);
 }
